// src/app/models/enums.ts
export enum AuthLevel {
  Guest = 0,
  User = 1,
  Admin = 2,
}
export enum Category {
  BABIES = 1,
  GIRLS = 2,
}
export enum BabiesSizes {
  NewBorn = '0-6',
  SixMonth = '6-12',
  OneYear = '12-18',
  EighteenMonth = '18-24',
  TwoYears = '24-36',
}
export enum GirlsSizes {
  Four = '4',
  Six = '6',
  Eight = '8',
  Ten = '10',
  Twelve = '12',
  Fourteen = '14',
  Sixteen = '16',
  Eighteen = '18',
}